查询新鲜事时合并redis和数据库的点赞数

+++ fresh-system/src/main/java/com/creamakers/fresh/system/domain/vo/response/FreshNewsResp.java
@@ -35,7 +35,7 @@ public class FreshNewsResp {
     private String tags;
 
     // 点赞数量
-    private Integer liked;
+    private Long liked;
 
     // 评论数量
     private Integer comments;
